+++ src/Shape.php
@@ -127,6 +127,16 @@ abstract class Shape
         return $draw;
     }
 
+    protected function transformRotate(ImagickDraw $draw, float $rotation, bool $invert = false): ImagickDraw
+    {
+        if ($invert) {
+            $rotation *= -1;
+        }
+
+        $draw->rotate($rotation);
+        return $draw;
+    }
+
     protected function transform(ImagickDraw $draw, bool $invert = false): ImagickDraw
     {
         $instructions = $this->transformInstructions();
@@ -145,6 +155,9 @@ abstract class Shape
                     break;
                 case 'scale':
                     $draw = $this->transformScale($draw, $arguments, $invert);
+                    break;
+                case 'rotate':
+                    $draw = $this->transformRotate($draw, $arguments[0], $invert);
                     break;
                 default:
                     throw new UnsupportedTransformException();